requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
async-lru==2.0.4
asyncio-atexit==1.0.1
betterproto==2.0.0b6
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Deprecated==1.2.18
dnspython==2.7.0
einops==0.8.0
email_validator==2.2.0
exceptiongroup==1.2.2
filelock==3.17.0
fonttools==4.55.8
fsspec==2025.2.0
googleapis-common-protos==1.66.0
grpclib==0.4.7
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.1.0
id==1.5.0
idna==3.10
importlib-resources==5.13.0
importlib_metadata==8.5.0
in-toto-attestation==0.9.3
Jinja2==3.1.5
kagglehub==0.3.7
kiwisolver==1.4.8
lpips==0.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
model-signing==0.2.0
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.11.0.86
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-httpx==0.51b0
opentelemetry-instrumentation-requests==0.51b0
opentelemetry-instrumentation-threading==0.51b0
opentelemetry-instrumentation-urllib==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
protobuf==5.29.3
psutil==5.9.8
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rfc3161-client==0.1.2
rfc8785==0.1.4
rich==13.9.4
scipy==1.15.1
securesystemslib==1.2.0
sigstore==3.6.1
sigstore-protobuf-specs==0.3.2
sigstore-rekor-types==0.0.18
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
synapseclient==4.7.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
tuf==5.1.0
typing_extensions==4.12.2
urllib3==1.26.20
wrapt==1.17.2
zipp==3.21.0
